% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_simex.R
\name{bootstrap_simex}
\alias{bootstrap_simex}
\title{Calculates SIMEX estimates for resampled data}
\usage{
bootstrap_simex(pred_mat, lambdavec, fail_times, fail_ind, init_b, B,
  simex_sd, n)
}
\arguments{
\item{pred_mat}{Matrix of predictors}

\item{fail_times}{Vector of error-prone
failure times}

\item{fail_ind}{Matrix of failure indicators}

\item{init_b}{Matrix of starting values
for optimization}

\item{B}{Parameter for SIMEX algorithm}

\item{simex_sd}{Standard deviation of
multiplicative error in event-time}

\item{n}{Number of samples}

\item{lambavec}{Vector of lambda params
for SIMEX algorithm}
}
\value{
Matrix of corrected betas from 
resampled data
}
\description{
Runs the bootstrap to calculate standard
errors for the SIMEX procedure
}
