% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simex_surv_outcome.R
\name{simex_surv_outcome}
\alias{simex_surv_outcome}
\title{Calculates SIMEX estimates for error in a time-to-event}
\usage{
simex_surv_outcome(pred_mat, lambdavec, fail_times, fail_ind, B = 50,
  simex_sd)
}
\arguments{
\item{pred_mat}{Matrix of predictors}

\item{fail_times}{Vector of error-prone
failure times}

\item{fail_ind}{Matrix of failure indicators}

\item{B}{Parameter for SIMEX algorithm}

\item{simex_sd}{Standard deviation of
multiplicative error in event-time}

\item{lambavec}{Vector of lambda params
for SIMEX algorithm}
}
\value{
Matrix of corrected betas
}
\description{
The SIMEX method adds additional error to the error-prone
time-to-event that has either known standard deviation or
is calculated from a validation subset. Then, a quadratic
function is fit to the error-prone estimates with
additional error and we extrapolate the function to
where theoretically, there is no error
}
